+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed WebFinger lookup not working on Yggdrasil addresses.
+- Display followers-only reposts of viewer's posts in profile timeline.
 
 ## [4.3.1] - 2025-05-18